Changed FreeAbelian to use int instead of long int.

+++ sources/Atlas.h
@@ -165,11 +165,11 @@ namespace atlas {
   using hashtable::HashTable;
 
   namespace free_abelian {
-    template<typename T, typename C=long int, typename Compare=std::less<T> >
+    template<typename T, typename C=int, typename Compare=std::less<T> >
       struct Free_Abelian;
-    template<typename T, typename C=long int, typename Compare=std::less<T> >
+    template<typename T, typename C=int, typename Compare=std::less<T> >
       struct Monoid_Ring;
-    template<typename T, typename C=long int, typename Compare=std::less<T> >
+    template<typename T, typename C=int, typename Compare=std::less<T> >
       class Free_Abelian_light;
   }
   using free_abelian::Free_Abelian;
