+++ sources/utilities/free_abelian_fwd.h
@@ -16,11 +16,11 @@ namespace atlas {
 
 namespace free_abelian {
 
-template<typename T, typename C=long int, typename Compare=std::less<T> >
+template<typename T, typename C=int, typename Compare=std::less<T> >
   struct Free_Abelian;
-template<typename T, typename C=long int, typename Compare=std::less<T> >
+template<typename T, typename C=int, typename Compare=std::less<T> >
   struct Monoid_Ring;
-template<typename T, typename C=long int, typename Compare=std::less<T> >
+template<typename T, typename C=int, typename Compare=std::less<T> >
   class Free_Abelian_light;
 }
